auth

+++ src/app/api/character/[id]/route.ts
@@ -1,27 +1,38 @@
 import { NextRequest, NextResponse } from "next/server";
 import prisma from "@/app/prisma/prisma";
 import { z } from "zod";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../../auth/[...nextauth]/auth";
 
-async function GET(
-  request: NextRequest,
-  { params }: { params: { id: string } }
-) {
-  const userId = "wll";
-  const id = parseInt(params.id, 10);
-  const character = await prisma.character.findFirst({
-    where: { id, userId, deletedAt: { equals: null } },
-  });
-  return NextResponse.json({ data: { character } });
-}
+// async function GET(
+//   request: NextRequest,
+//   { params }: { params: { id: string } }
+// ) {
+//   const session = await getServerSession(authOptions);
+//   if (!session?.user) {
+//     return NextResponse.json({}, { status: 401 });
+//   }
+//   const userId = session.user.id;
+//   const id = parseInt(params.id, 10);
+//   const character = await prisma.character.findFirst({
+//     where: { id, userId, deletedAt: { equals: null } },
+//   });
+//   return NextResponse.json({ data: { character } });
+// }
 
 async function DELETE(
   request: NextRequest,
   { params }: { params: { id: string } }
 ) {
+  const session = await getServerSession(authOptions);
+  if (!session?.user) {
+    return NextResponse.json({}, { status: 401 });
+  }
+  const userId = session.user.id;
   try {
     const id = parseInt(params.id, 10);
     await prisma.character.delete({
-      where: { id },
+      where: { id, userId },
     });
   } catch (error) {
     return NextResponse.error();
@@ -38,12 +49,17 @@ async function PUT(
   request: NextRequest,
   { params }: { params: { id: string } }
 ) {
+  const session = await getServerSession(authOptions);
+  if (!session?.user) {
+    return NextResponse.json({}, { status: 401 });
+  }
+  const userId = session.user.id;
   const id = parseInt(params.id, 10);
   const data = await request.json();
   try {
     const params = updateInput.parse(data);
     await prisma.character.update({
-      where: { id },
+      where: { id, userId },
       data: {
         remark: params.remark,
         name: params.name,
@@ -59,4 +75,4 @@ async function PUT(
   }
 }
 
-export { DELETE, PUT, GET };
+export { DELETE, PUT };
